# docs/api/apex-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "ConsumerRead": {
        "properties": {
          "consumer_type": {
            "$ref": "#/components/schemas/ConsumerType"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key_fingerprint": {
            "description": "First 8 hex chars of the stored key hash (NOT of the raw key, which is never persisted). Stable identifier for a credential.",
            "title": "Key Fingerprint",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "rotated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotated At"
          },
          "rotation_count": {
            "default": 0,
            "title": "Rotation Count",
            "type": "integer"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/ScopeRef"
            },
            "title": "Scopes",
            "type": "array"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "required": [
          "id",
          "name",
          "consumer_type",
          "role",
          "enabled",
          "scopes",
          "created_at",
          "last_used_at",
          "key_fingerprint"
        ],
        "title": "ConsumerRead",
        "type": "object"
      },
      "ConsumerType": {
        "enum": [
          "dashboard",
          "headless",
          "internal"
        ],
        "title": "ConsumerType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Role": {
        "description": "Ordered roles: viewer < operator < admin.",
        "enum": [
          "viewer",
          "operator",
          "admin"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "RotateConsumerKeyRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "grace_period_seconds": {
            "default": 0,
            "maximum": 604800.0,
            "minimum": 0.0,
            "title": "Grace Period Seconds",
            "type": "integer"
          }
        },
        "title": "RotateConsumerKeyRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "APEX Orchestration Engine \u2014 Domain API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/consumers": {
      "get": {
        "operationId": "listConsumers",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsumerRead"
                  },
                  "title": "Response Listconsumers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Consumers",
        "tags": [
          "consumers"
        ]
      },
      "post": {
        "operationId": "createConsumer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Consumer",
        "tags": [
          "consumers"
        ]
      }
    },
    "/v1/admin/consumers/{consumer_id}": {
      "delete": {
        "operationId": "deleteConsumer",
        "parameters": [
          {
            "description": "Consumer id",
            "in": "path",
            "name": "consumer_id",
            "required": true,
            "schema": {
              "description": "Consumer id",
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[^\\x00]*$",
              "title": "Consumer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Consumer",
        "tags": [
          "consumers"
        ]
      },
      "get": {
        "operationId": "getConsumer",
        "parameters": [
          {
            "description": "Consumer id",
            "in": "path",
            "name": "consumer_id",
            "required": true,
            "schema": {
              "description": "Consumer id",
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[^\\x00]*$",
              "title": "Consumer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Consumer",
        "tags": [
          "consumers"
        ]
      },
      "patch": {
        "operationId": "updateConsumer",
        "parameters": [
          {
            "description": "Consumer id",
            "in": "path",
            "name": "consumer_id",
            "required": true,
            "schema": {
              "description": "Consumer id",
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[^\\x00]*$",
              "title": "Consumer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumerUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Consumer",
        "tags": [
          "consumers"
        ]
      }
    },
    "/v1/admin/consumers/{consumer_id}/rotate": {
      "post": {
        "operationId": "rotateConsumerKey",
        "parameters": [
          {
            "description": "Consumer id",
            "in": "path",
            "name": "consumer_id",
            "required": true,
            "schema": {
              "description": "Consumer id",
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[^\\x00]*$",
              "title": "Consumer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RotateConsumerKeyRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Consumer Key",
        "tags": [
          "consumers"
        ]
      }
    },
...
  }
}